organised item loading content, updating and drawing with a list.
half way to making potion work with events. need to first try and make it heal with events only, and then make player able to collide and collect it, and use hotkey to heal.

still need to work on scene loading, unloading and reseting.

Also need to fix issue that skips intro scene when you get a game over